# Conflicts:
#	Source/Group3_PetShop/src/main/java/com/t3h/group3_petshop/service/IProductService.java
#	Source/Group3_PetShop/src/main/java/com/t3h/group3_petshop/service/impl/CartServiceImpl.java
#	Source/Group3_PetShop/src/main/java/com/t3h/group3_petshop/service/impl/ProductServiceImpl.java
#	Source/Group3_PetShop/src/main/resources/templates/layout/header.html

+++ Source/Group3_PetShop/src/main/java/com/t3h/group3_petshop/model/request/ProductFilterRequest.java
@@ -7,6 +7,6 @@ public class ProductFilterRequest {
     private String name;
     private String code;
     private Float price;
-    private Integer sizeId;
+    private Long sizeId;
     private Integer categoryId ;
 }
